+++ src/main/java/com/ojs/service/content/v1/dto/Journal.java
@@ -6,18 +6,18 @@ import org.springframework.hateoas.ResourceSupport;
  * Created by agile on 6/23/17.
  */
 public class Journal extends ResourceSupport {
-    private final int journalId;
+    private final long journalId;
     private final String path;
     private String description;
 
 
-    public Journal(int journalId, String path) {
+    public Journal(Long journalId, String path) {
         this.journalId = journalId;
         this.path = path;
 
     }
 
-    public int getJournalId() {
+    public long getJournalId() {
         return journalId;
     }
